pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watermark-analysis"
version = "0.1.0"
description = "Watermark encoders, removal attacks, and evaluation metrics."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "MBZUAI Watermark-Analysis Team" }]
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "diffusers>=0.25",
    "transformers>=4.35",
    "onnxruntime>=1.16",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "tqdm>=4.65",
    "wandb>=0.16",
    "scikit-learn>=1.3",
    "scikit-image>=0.21",
    "scipy>=1.11",
    "lpips>=0.1.4",
    "clean-fid>=0.1.35",
    "pytorch-fid>=0.3",
    "pywavelets>=1.4",
    "compressai>=1.2",
    "matplotlib>=3.7",
]

[project.scripts]
wm-generate-prompts = "watermark_analysis.cli:generate_prompts_main"
wm-create-dataset = "watermark_analysis.cli:create_dataset_main"
wm-train-adaptive = "watermark_analysis.cli:train_adaptive_main"
wm-apply-adaptive = "watermark_analysis.cli:apply_adaptive_main"
wm-eval-performance = "watermark_analysis.cli:eval_performance_main"
wm-eval-quality = "watermark_analysis.cli:eval_quality_main"
wm-inspect-onnx = "watermark_analysis.cli:inspect_onnx_main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
